refactor: domain/vocabulary package 재배치
