+++ server/api/src/main/java/vook/server/api/domain/vocabulary/logic/vocabulary/VocabularyCreateCommand.java
@@ -1,10 +1,10 @@
-package vook.server.api.domain.vocabulary.logic.dto;
+package vook.server.api.domain.vocabulary.logic.vocabulary;
 
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.Size;
 import lombok.Builder;
-import vook.server.api.domain.vocabulary.model.UserUid;
+import vook.server.api.domain.vocabulary.model.vocabulary.UserUid;
 
 @Builder
 public record VocabularyCreateCommand(
