+++ server/api/src/main/java/vook/server/api/domain/vocabulary/model/vocabulary/UserUid.java
@@ -1,4 +1,4 @@
-package vook.server.api.domain.vocabulary.model;
+package vook.server.api.domain.vocabulary.model.vocabulary;
 
 import lombok.AllArgsConstructor;
 import lombok.EqualsAndHashCode;
