+++ server/api/src/main/java/vook/server/api/web/term/usecase/UpdateTermUseCase.java
@@ -2,9 +2,9 @@ package vook.server.api.web.term.usecase;
 
 import lombok.RequiredArgsConstructor;
 import vook.server.api.domain.user.logic.UserLogic;
-import vook.server.api.domain.vocabulary.logic.TermLogic;
-import vook.server.api.domain.vocabulary.logic.dto.TermUpdateCommand;
-import vook.server.api.domain.vocabulary.model.Term;
+import vook.server.api.domain.vocabulary.logic.term.TermLogic;
+import vook.server.api.domain.vocabulary.logic.term.TermUpdateCommand;
+import vook.server.api.domain.vocabulary.model.term.Term;
 import vook.server.api.globalcommon.annotation.UseCase;
 import vook.server.api.policy.VocabularyPolicy;
