+++ server/api/src/main/java/vook/server/api/web/vocabulary/usecase/CreateVocabularyUseCase.java
@@ -2,9 +2,9 @@ package vook.server.api.web.vocabulary.usecase;
 
 import lombok.RequiredArgsConstructor;
 import vook.server.api.domain.user.logic.UserLogic;
-import vook.server.api.domain.vocabulary.logic.VocabularyLogic;
-import vook.server.api.domain.vocabulary.logic.dto.VocabularyCreateCommand;
-import vook.server.api.domain.vocabulary.model.UserUid;
+import vook.server.api.domain.vocabulary.logic.vocabulary.VocabularyCreateCommand;
+import vook.server.api.domain.vocabulary.logic.vocabulary.VocabularyLogic;
+import vook.server.api.domain.vocabulary.model.vocabulary.UserUid;
 import vook.server.api.globalcommon.annotation.UseCase;
 
 @UseCase
