+++ server/api/src/main/java/vook/server/api/web/vocabulary/usecase/UpdateVocabularyUseCase.java
@@ -2,9 +2,9 @@ package vook.server.api.web.vocabulary.usecase;
 
 import lombok.RequiredArgsConstructor;
 import vook.server.api.domain.user.logic.UserLogic;
-import vook.server.api.domain.vocabulary.logic.VocabularyLogic;
-import vook.server.api.domain.vocabulary.logic.dto.VocabularyUpdateCommand;
-import vook.server.api.domain.vocabulary.model.Vocabulary;
+import vook.server.api.domain.vocabulary.logic.vocabulary.VocabularyLogic;
+import vook.server.api.domain.vocabulary.logic.vocabulary.VocabularyUpdateCommand;
+import vook.server.api.domain.vocabulary.model.vocabulary.Vocabulary;
 import vook.server.api.globalcommon.annotation.UseCase;
 import vook.server.api.policy.VocabularyPolicy;
