+++ server/api/src/test/java/vook/server/api/infra/vocabulary/DefaultVocabularyRepositoryTest.java
@@ -5,8 +5,8 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.transaction.annotation.Transactional;
-import vook.server.api.domain.vocabulary.model.UserUid;
-import vook.server.api.domain.vocabulary.model.Vocabulary;
+import vook.server.api.domain.vocabulary.model.vocabulary.UserUid;
+import vook.server.api.domain.vocabulary.model.vocabulary.Vocabulary;
 import vook.server.api.infra.vocabulary.cache.UserVocabularyCache;
 import vook.server.api.infra.vocabulary.cache.UserVocabularyCacheRepository;
 import vook.server.api.testhelper.IntegrationTestBase;
