+++ server/api/src/test/java/vook/server/api/testhelper/creator/TestVocabularyCreator.java
@@ -4,14 +4,14 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
 import org.springframework.transaction.annotation.Transactional;
 import vook.server.api.domain.user.model.User;
-import vook.server.api.domain.vocabulary.logic.TermLogic;
-import vook.server.api.domain.vocabulary.logic.VocabularyLogic;
-import vook.server.api.domain.vocabulary.logic.dto.TermCreateAllCommand;
-import vook.server.api.domain.vocabulary.logic.dto.TermCreateCommand;
-import vook.server.api.domain.vocabulary.logic.dto.VocabularyCreateCommand;
-import vook.server.api.domain.vocabulary.model.Term;
-import vook.server.api.domain.vocabulary.model.UserUid;
-import vook.server.api.domain.vocabulary.model.Vocabulary;
+import vook.server.api.domain.vocabulary.logic.term.TermCreateAllCommand;
+import vook.server.api.domain.vocabulary.logic.term.TermCreateCommand;
+import vook.server.api.domain.vocabulary.logic.term.TermLogic;
+import vook.server.api.domain.vocabulary.logic.vocabulary.VocabularyCreateCommand;
+import vook.server.api.domain.vocabulary.logic.vocabulary.VocabularyLogic;
+import vook.server.api.domain.vocabulary.model.term.Term;
+import vook.server.api.domain.vocabulary.model.vocabulary.UserUid;
+import vook.server.api.domain.vocabulary.model.vocabulary.Vocabulary;
 
 import java.util.List;
 import java.util.concurrent.atomic.AtomicInteger;
