+++ server/api/src/test/java/vook/server/api/web/user/usecase/OnboardingUserUseCaseTest.java
@@ -8,9 +8,9 @@ import vook.server.api.domain.template_vocabulary.model.TemplateVocabularyType;
 import vook.server.api.domain.user.model.Funnel;
 import vook.server.api.domain.user.model.Job;
 import vook.server.api.domain.user.model.UserRepository;
-import vook.server.api.domain.vocabulary.model.TermRepository;
-import vook.server.api.domain.vocabulary.model.UserUid;
-import vook.server.api.domain.vocabulary.model.VocabularyRepository;
+import vook.server.api.domain.vocabulary.model.term.TermRepository;
+import vook.server.api.domain.vocabulary.model.vocabulary.UserUid;
+import vook.server.api.domain.vocabulary.model.vocabulary.VocabularyRepository;
 import vook.server.api.testhelper.IntegrationTestBase;
 import vook.server.api.testhelper.creator.TestTemplateVocabularyCreator;
 import vook.server.api.testhelper.creator.TestUserCreator;
